+++ src/main/java/com/github/johnmedlockdev/project1john/business/SparkConfig.java
@@ -1,7 +1,10 @@
 package com.github.johnmedlockdev.project1john.business;
 
+import com.github.johnmedlockdev.project1john.controllers.RestfulController;
 import org.apache.spark.SparkConf;
 import org.apache.spark.api.java.JavaSparkContext;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;