+++ src/main/java/com/github/johnmedlockdev/project1john/controllers/StaticController.java
@@ -2,6 +2,8 @@ package com.github.johnmedlockdev.project1john.controllers;
 
 import com.github.johnmedlockdev.project1john.repositories.SparkRepository;
 import com.github.johnmedlockdev.project1john.utils.LoadFile;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
@@ -13,6 +15,8 @@ import org.springframework.web.multipart.MultipartFile;
 @Controller
 public class StaticController {
 
+    Logger logger = LoggerFactory.getLogger(StaticController.class);
+
     @Autowired
     LoadFile loadFile;
 
@@ -21,12 +25,14 @@ public class StaticController {
 
     @RequestMapping(value = "/")
     public String index() {
+        logger.info("Request @ /");
         sparkRepository.deleteAll();
         return "index";
     }
 
     @RequestMapping(value = "/graph")
     public String graph() {
+        logger.info("Request @ /graph");
         return "graph";
     }