+++ src/main/java/com/github/johnmedlockdev/project1john/utils/LoadFile.java
@@ -1,6 +1,9 @@
 package com.github.johnmedlockdev.project1john.utils;
 
 import com.github.johnmedlockdev.project1john.business.SparkProcess;
+import org.slf4j.ILoggerFactory;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
@@ -13,6 +16,7 @@ import java.nio.file.Paths;
 @Component
 public class LoadFile {
 
+    Logger logger = LoggerFactory.getLogger(LoadFile.class);
     @Autowired
     SparkProcess sparkProcess;
 
@@ -23,9 +27,9 @@ public class LoadFile {
         try (BufferedWriter w = Files.newBufferedWriter(Paths.get(fileName))) {
             w.write(new String(multipartFile.getBytes()));
         } catch (IOException ioException) {
+            logger.info("Something went wrong with io");
             ioException.printStackTrace();
-            // TODO: 5/13/2020
         }
-        sparkProcess.spark(fileName);
+        sparkProcess.init(fileName);
     }
 }